feat: move plotting methods into `Column`
