+++ src/safeds/plotting/__init__.py
@@ -1,5 +1,3 @@
-from ._boxplot import boxplot
 from ._correlation_heatmap import correlation_heatmap
-from ._histogram import histogram
 from ._lineplot import lineplot
 from ._scatterplot import scatterplot
